move drive subsystem out of its special folder
