+++ src/main/java/frc/robot/commands/ArcadeDrive.java
@@ -2,7 +2,7 @@ package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.controllers.DriverController;
-import frc.robot.subsystems.drive.DiffDriveSubsystem;
+import frc.robot.subsystems.DiffDriveSubsystem;
 
 public class ArcadeDrive extends CommandBase {
     private final DiffDriveSubsystem diffDriveSubsystem = DiffDriveSubsystem.getInstance();
