+++ src/main/java/frc/robot/commands/SyncLimelightPose.java
@@ -15,8 +15,8 @@ import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import edu.wpi.first.wpilibj2.command.Subsystem;
+import frc.robot.subsystems.DiffDriveSubsystem;
 import frc.robot.subsystems.LimelightSubsytem;
-import frc.robot.subsystems.drive.DiffDriveSubsystem;
 
 public class SyncLimelightPose extends CommandBase {
   private final LimelightSubsytem limelightSubsytem;
