+++ src/main/java/frc/robot/commands/waypointCommand.java
@@ -10,8 +10,8 @@ import java.util.Set;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.Subsystem;
+import frc.robot.subsystems.DiffDriveSubsystem;
 import frc.robot.subsystems.LimelightSubsytem;
-import frc.robot.subsystems.drive.DiffDriveSubsystem;
 import frc.robot.Robot;
 import frc.robot.RobotState;
